python/requirements.txt
ijson==2.3
python-dateutil==2.8.0
pytz==2018.9
PyXB==1.2.6
six==1.12.0
psycopg2==2.8.2
typing==3.7.4
